Add option to accept or reject task invitations

+++ src/components/TaskDisplay/TaskDisplay.tsx
@@ -4,7 +4,7 @@ import OutsideAlerter from '../OutsideAlerter/OutsideAlerter';
 import { Task } from '../../types/types';
 import { format, parseISO } from 'date-fns';
 import { es } from 'date-fns/locale';
-import { deleteDoc, doc } from 'firebase/firestore';
+import { deleteDoc, doc, getDoc, updateDoc } from 'firebase/firestore';
 import { db } from '../../firebase';
 import ParticipantAvatar from '../ParticipantAvatar/ParticipantAvatar';
 
@@ -16,10 +16,63 @@ type Props = {
 export default function TaskDisplay({ task, close }: Props) {
   const user = useContext(AuthContext);
 
+  // Conditions to select elements to display
+  const isCreator = task.creator.id === user.id;
+  const userDataInParticipantsArr = task.participants.find(
+    (person) => person.id === user.id
+  );
+  const hasAnswered =
+    !!userDataInParticipantsArr && !!userDataInParticipantsArr.isAccepted;
+
   async function handleDelete() {
+    if (isCreator) {
+      try {
+        close();
+        await deleteDoc(doc(db, 'users', user.id, 'tasks', task.id));
+      } catch (err) {
+        console.log(err);
+      }
+    } else {
+      const docRef = doc(db, 'users', task.creator.id, 'tasks', task.id);
+      try {
+        close();
+        const res = await getDoc(docRef);
+        if (res.exists()) {
+          const data = res.data() || [];
+          const participantsArr = [
+            ...data.participants.map((person: { id: string }) =>
+              person.id === user.id ? { ...person, isAccepted: false } : person
+            ),
+          ];
+          await updateDoc(docRef, {
+            participants: participantsArr,
+          });
+        }
+      } catch (err) {
+        console.log(err);
+      }
+    }
+  }
+
+  async function handleAnswer(isAccepted: boolean = true) {
+    const docRef = doc(db, 'users', task.creator.id, 'tasks', task.id);
+
     try {
       close();
-      await deleteDoc(doc(db, 'users', user.id, 'tasks', task.id));
+      const res = await getDoc(docRef);
+      if (res.exists()) {
+        const data = res.data() || [];
+        const participantsArr = [
+          ...data.participants.map((person: { id: string }) =>
+            person.id === user.id
+              ? { ...person, isAccepted: isAccepted }
+              : person
+          ),
+        ];
+        await updateDoc(docRef, {
+          participants: participantsArr,
+        });
+      }
     } catch (err) {
       console.log(err);
     }
@@ -72,26 +125,46 @@ export default function TaskDisplay({ task, close }: Props) {
           </div>
 
           <div className="pt-4">
-            <button
-              onClick={handleDelete}
-              className="flex items-center gap-1 font-bold text-emerald-700 dark:text-emerald-500"
-            >
-              ELIMINAR{' '}
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 24 24"
-                strokeWidth={1.5}
-                stroke="currentColor"
-                className="h-4 w-4"
+            {!isCreator && !hasAnswered ? (
+              <div>
+                Has sido invitado a esta actividad
+                <div className="mt-2 flex gap-4">
+                  <button
+                    onClick={() => handleAnswer(true)}
+                    className="flex items-center gap-1 font-bold text-emerald-700 dark:text-emerald-500"
+                  >
+                    Aceptar
+                  </button>
+                  <button
+                    onClick={() => handleAnswer(false)}
+                    className="flex items-center gap-1 font-bold text-red-700 dark:text-red-500"
+                  >
+                    Rechazar
+                  </button>
+                </div>
+              </div>
+            ) : (
+              <button
+                onClick={handleDelete}
+                className="flex items-center gap-1 font-bold text-emerald-700 dark:text-emerald-500"
               >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"
-                />
-              </svg>
-            </button>
+                {isCreator ? 'ELIMINAR ' : 'ELIMINAR PARA MÍ '}
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  strokeWidth={1.5}
+                  stroke="currentColor"
+                  className="h-4 w-4"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"
+                  />
+                </svg>
+              </button>
+            )}
           </div>
         </div>
       </OutsideAlerter>
